NexCompare/requirements.txt
PyYAML==6.0.1
json-diff==1.0.1
